codes/requirements.txt
numpy
opencv-python
pyyaml
future
tensorboard
prefetch_generator
# conda install pytorch torchvision cudatoolkit=10.2 -c pytorch